adjust packages
