+++ studies/dyn4j_for_sim/src/main/java/org/team100/sim/Body100.java
@@ -1,4 +1,4 @@
-package frc.robot;
+package org.team100.sim;
 
 import java.util.List;
