+++ studies/dyn4j_for_sim/src/main/java/org/team100/sim/FixedFilter.java
@@ -1,4 +1,4 @@
-package frc.robot;
+package org.team100.sim;
 
 import org.dyn4j.collision.TypeFilter;
