+++ studies/dyn4j_for_sim/src/main/java/org/team100/sim/Friend.java
@@ -1,4 +1,4 @@
-package frc.robot;
+package org.team100.sim;
 
 import org.dyn4j.geometry.Vector2;
