+++ studies/dyn4j_for_sim/src/main/java/org/team100/sim/Main.java
@@ -1,6 +1,7 @@
-package frc.robot;
+package org.team100.sim;
 
 import edu.wpi.first.wpilibj.RobotBase;
+import frc.robot.Robot;
 
 public final class Main {
   private Main() {}
