+++ studies/dyn4j_for_sim/src/main/java/org/team100/sim/NoteFilter.java
@@ -1,4 +1,4 @@
-package frc.robot;
+package org.team100.sim;
 
 /**
  * Notes collide with each other and with fixed objects, but robots drive over
