+++ studies/dyn4j_for_sim/src/main/java/org/team100/sim/RobotFilter.java
@@ -1,4 +1,4 @@
-package frc.robot;
+package org.team100.sim;
 
 /**
  * Robots collide with each other and with fixed objects, but not notes.
